用户登录状态判断方式更改
原来是将用户登录后的信息保存在 redux 中，现在则是保存在 sessionStorage 中。这样能够避免在刷新页面的时候，虽然 cookies 没有过期，但是由于 redux 中的数据被清空而强制转跳到登陆页面的情况。
