+++ src/app.js
@@ -2,33 +2,30 @@ import React from 'react';
 import { Route, Switch } from 'react-router-dom';
 import { hot } from 'react-hot-loader/root';
 
+import { userDataEventListener } from 'util/custom-event';
 import { getUserDataStorage } from 'util/storege';
-import store from './store.js';
-import Login from 'page/login/index.js';
+import Login from 'page/login/index';
 import PageRoute from 'page/index';
 
 class App extends React.Component {
   state = {
-    userData: {}
+    userData: getUserDataStorage()
   }
-
-  componentWillMount() {
-    const userData = store.getState().login.data;
-    this.setState({ userData });
-  }
-
   componentDidMount() {
-    store.subscribe(() => {
-      const userData = store.getState().login.data;
-      this.setState({ userData });
+    userDataEventListener(this.handleUserDataChange);
+  }
+
+  /**
+   * @param {string} action login or logout
+   */
+  handleUserDataChange = (action) => {
+    this.setState({
+      userData: getUserDataStorage()
     });
   }
 
   handleRender = ({ location }) => {
-    const { userData } = this.state;
-    // const userData = getUserDataStorage();
-    const keys = Object.keys(userData);
-    return keys.length === 0 ? <Login /> : <PageRoute location={location} />;
+    return this.state.userData ? <PageRoute location={location} /> : <Login />;
   }
   render() {
     return (
