--- src/page/login/actionTypes.js
@@ -1,2 +1,0 @@
-export const DO_LOG_IN = 'DO_LOG_IN';
-export const DO_LOG_OUT = 'DO_LOG_OUT';
