--- src/page/login/view/container.js
@@ -1,15 +1,0 @@
-import { connect } from 'react-redux';
-import { withRouter } from 'react-router-dom';
-
-import * as actions from '../actions.js';
-import Presentation from './presentation.js';
-
-const mapStateToProps = state => ({
-  login: state.login,
-});
-
-const mapDispatchToProps = {
-  ...actions
-};
-
-export default withRouter(connect(mapStateToProps, mapDispatchToProps)(Presentation));
