+++ src/page/reducer.js
@@ -1,4 +1,3 @@
-import { reducer as LoginReducer } from './login/index';
 import { reducer as HomeReducer } from './home/index';
 import { reducer as userReducer } from './user/index';
 import { reducer as productReducer } from './product/index';
@@ -6,7 +5,6 @@ import { reducer as categoryReducer } from './category/index';
 import { reducer as orderReducer } from './order';
 
 const reducer = {
-  login: LoginReducer,
   home: HomeReducer,
   ...userReducer,
   ...productReducer,
